Converted DelegationmetadataService to EF

+++ src/apps/Altinn.AccessManagement/src/Altinn.AccessManagement.Core/Models/InstanceDelegationChange.cs
@@ -13,7 +13,7 @@ namespace Altinn.AccessManagement.Core.Models
         /// Gets or sets the instance delegation change id
         /// </summary>
         [JsonPropertyName("instancedelegationchangeid")]
-        public int InstanceDelegationChangeId { get; set; }
+        public long InstanceDelegationChangeId { get; set; }
 
         /// <summary>
         /// Gets or sets the DelegationChangeType.
